requirements.txt
Django>=4.2.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
djangorestframework>=3.14.0
python-decouple>=3.8

